+++ src/rules/relative-font-units.ts
@@ -1,7 +1,7 @@
-import css from '@eslint/css';
+import { cssRules } from '../utils/css';
 
 /**
  * Wrapper for \@eslint/css relative-font-units rule
  * Prefers rem/em over px for better accessibility
  */
-export const relativeFontUnits = css.rules['relative-font-units'];
+export const relativeFontUnits = cssRules['relative-font-units'];