+++ src/rules/use-layers.ts
@@ -1,7 +1,7 @@
-import css from '@eslint/css';
+import { cssRules } from '../utils/css';
 
 /**
  * Wrapper for \@eslint/css use-layers rule
  * Encourages use of `@layer` for better CSS architecture
  */
-export const useLayers = css.rules['use-layers'];
+export const useLayers = cssRules['use-layers'];